% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution}
\alias{contribution}
\title{Compute Individual and Cumulative Contribution Ratios}
\usage{
contribution(D)
}
\arguments{
\item{D}{A numeric vector representing singular values or eigenvalues.}
}
\value{
A list containing:
\describe{
  \item{individual_contribution}{A numeric vector of individual contribution ratios.}
  \item{cumulative_contribution}{A numeric vector of cumulative contribution ratios.}
}
}
\description{
Calculates the individual and cumulative contribution ratios based on the provided vector of singular values or eigenvalues.
}
