% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold.R
\name{mode1unfold}
\alias{mode1unfold}
\title{Mode-1 Unfolding of a Tensor}
\usage{
mode1unfold(K)
}
\arguments{
\item{K}{A tensor, typically generated by the `cumulant_tensor_cpp` function.}
}
\value{
A matrix representing the tensor unfolded along the first mode.
}
\description{
Unfolds a tensor along the first mode into a matrix.
}
