% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_reconstruction.R
\name{reconstruct}
\alias{reconstruct}
\title{Linear reconstruction of data}
\usage{
reconstruct(Score, Uk, col_means = NULL, col_sds = NULL)
}
\arguments{
\item{Score}{A numeric matrix of principal component scores, typically obtained from the `scores` function.}

\item{Uk}{A matrix of the top `k` left singular vectors (loadings) obtained from the `reduction` function.}

\item{col_means}{A numeric vector containing the means of the original data columns, obtained from the `scale` function.}

\item{col_sds}{A numeric vector containing the standard deviations of the original data columns, obtained from the `scale` function.}
}
\value{
A numeric matrix representing the reconstructed data using linear reconstruction.
}
\description{
Reconstructs the original data from the principal component scores and loadings using a linear method.
}
