% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{reconstruction_plot}
\alias{reconstruction_plot}
\title{Plot Original vs. Reconstructed Data}
\usage{
reconstruction_plot(df, result, k = 1)
}
\arguments{
\item{df}{A matrix or data frame containing the original data, where rows represent observations and columns represent variables.}

\item{result}{the result of cokpca function containing:
\describe{
  \item{reconstructed_data}{A matrix or data frame containing the reconstructed data, matching the structure of \code{df}.}
}}

\item{k}{An integer specifying which variable's data to plot (i.e., which column in the data matrices). Default is \code{1}.}
}
\value{
None. This function generates a line plot as a side effect.
}
\description{
Visualizes the comparison between original data and its reconstruction for a specified variable.
}
\details{
The function plots the original and reconstructed data for the specified variable \code{k}. It displays the original data in blue and the reconstructed data in red, with a legend indicating each.
}
\examples{
data = matrix(rnorm(100), nrow = 20, ncol = 5)
result = cokpca(data)
reconstruction_plot(df = data, result = result, k = 2)

}
