% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduction}
\alias{reduction}
\title{Dimensionality Reduction}
\usage{
reduction(U, D, V, k = NULL, alpha = NULL)
}
\arguments{
\item{U}{A matrix of left singular vectors obtained from Singular Vector Decomposition.}

\item{D}{A numeric vector of singular values obtained from Singular Vector Decomposition.}

\item{V}{A matrix of right singular vectors obtained from Singular Vector Decomposition.}

\item{k}{(Optional) An integer specifying the number of principal components to retain. Must be a positive integer no more than the length of `D`.}

\item{alpha}{(Optional) A numerical value in [0,1] specifying the cumulative contribution ratio to determine the number of principal components to retain.}
}
\value{
A list containing:
\describe{
  \item{Uk}{A matrix of the top `k` left singular vectors.}
  \item{Dk}{A diagonal matrix of the top `k` singular values.}
  \item{Vk}{A matrix of the top `k` right singular vectors.}
  \item{k}{The number of principal components selected.}
  \item{alpha}{The cumulative contribution ratio achieved with the selected `k` components.}
}
}
\description{
Reduces the dimension of the data.
}
