% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scores}
\alias{scores}
\title{Compute Scores for Principal Components}
\usage{
scores(data, Uk)
}
\arguments{
\item{data}{A numeric matrix of scaled data.}

\item{Uk}{A matrix of the top `k` left singular vectors obtained from the `reduction` function.}
}
\value{
A numeric matrix of scores.
}
\description{
Calculates the scores for selected principal components.
}
