% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold.R
\name{unfold}
\alias{unfold}
\title{Tensor Unfolding}
\usage{
unfold(K, row_idx, col_idx)
}
\arguments{
\item{K}{A tensor, typically generated by the `cumulant_tensor_cpp` function.}

\item{row_idx}{An integer vector specifying dimensions of the tensors to use as the row of the unfolded matrix.}

\item{col_idx}{An integer vector specifying dimensions of the tensors to use as the column of the unfolded matrix.}
}
\value{
A matrix representing the unfolded tensor.
}
\description{
Unfolds a tensor along specified modes into a matrix.
}
